# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_stacyvm-dev_stacyvm_internal_httputil.APIError:
    properties:
      code:
        $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.ErrorCode'
      message:
        type: string
    type: object
  github_com_stacyvm-dev_stacyvm_internal_httputil.ErrorCode:
    enum:
    - NOT_FOUND
    - BAD_REQUEST
    - INTERNAL_ERROR
    - UNAUTHORIZED
    - CONFLICT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeBadRequest
    - CodeInternal
    - CodeUnauth
    - CodeConflict
    - CodeUnavailable
...
  github_com_stacyvm-dev_stacyvm_internal_orchestrator.Sandbox:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      image:
        type: string
      memory_mb:
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      provider:
        type: string
      state:
        $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.SandboxState'
      vcpus:
        type: integer
    type: object
  github_com_stacyvm-dev_stacyvm_internal_orchestrator.SandboxState:
    enum:
    - creating
    - running
    - idle
    - destroyed
    - error
    type: string
    x-enum-varnames:
    - StateCreating
    - StateRunning
    - StateIdle
    - StateDestroyed
    - StateError
  github_com_stacyvm-dev_stacyvm_internal_orchestrator.SecretConfig:
    properties:
      inject_at:
        type: string
      name:
        type: string
    type: object
  github_com_stacyvm-dev_stacyvm_internal_orchestrator.SpawnRequest:
    properties:
      image:
        type: string
      memory_mb:
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      provider:
        type: string
      template:
        type: string
      ttl:
        type: string
      vcpus:
        type: integer
    type: object
  github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template:
    properties:
      allowed_hosts:
        items:
          type: string
        type: array
      cpu_cores:
        type: integer
      created_at:
        type: string
      description:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      image:
        type: string
      memory_mb:
        type: integer
      name:
        type: string
      pool_size:
        type: integer
      secrets:
        items:
          $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.SecretConfig'
        type: array
      setup:
        items:
          type: string
        type: array
      ttl_seconds:
        type: integer
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  internal_api_routes.ProviderInfo:
    properties:
      default:
        example: true
        type: boolean
      healthy:
        example: true
        type: boolean
      name:
        example: firecracker
        type: string
    type: object
  internal_api_routes.PruneResponse:
    properties:
      pruned:
        example: 3
        type: integer
    type: object
  internal_api_routes.StatusResponse:
    properties:
      status:
        example: destroyed
        type: string
    type: object
  internal_api_routes.TemplateSpawnOverrides:
    properties:
      provider:
        example: firecracker
        type: string
      ttl:
        example: 30m
        type: string
    type: object
host: localhost:7423
info:
  contact: {}
  description: StacyVM microVM sandbox orchestrator API
  title: StacyVM API
  version: "1.0"
paths:
...
  /providers:
    get:
      description: Return all registered providers with health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_routes.ProviderInfo'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List providers
      tags:
      - providers
...
  /sandboxes:
    delete:
      description: Destroy all expired sandboxes and return the count
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_routes.PruneResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Prune sandboxes
      tags:
      - sandboxes
    get:
      description: Return all active sandboxes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Sandbox'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: List sandboxes
      tags:
      - sandboxes
    post:
      consumes:
      - application/json
      description: Spawn a new sandbox with the given configuration
      parameters:
      - description: Spawn request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.SpawnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Sandbox'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create a sandbox
      tags:
      - sandboxes
...
  /templates:
    get:
      description: Return all registered templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: List templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Register a new sandbox template
      parameters:
      - description: Template definition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create a template
      tags:
      - templates
  /templates/{name}:
    delete:
      description: Delete a template by name
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_routes.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete a template
      tags:
      - templates
    get:
      description: Return a template by its name
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get a template
      tags:
      - templates
    put:
      consumes:
      - application/json
      description: Update an existing template by name
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      - description: Updated template
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Update a template
      tags:
      - templates
  /templates/{name}/spawn:
    post:
      consumes:
      - application/json
      description: Create a new sandbox using a template's configuration, with optional
        overrides
      parameters:
      - description: Template name
        in: path
        name: name
        required: true
        type: string
      - description: Optional overrides
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_routes.TemplateSpawnOverrides'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_orchestrator.Sandbox'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Spawn from template
      tags:
      - templates
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"